知识图谱建模/requirements.txt
altgraph==0.17.3
audioread==3.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.12.2
fsspec==2023.6.0
grpcio==1.57.0
grpcio-tools==1.57.0
huggingface-hub==0.16.4
idna==3.4
interchange==2021.0.4
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
libsndfile1==0.0.1
llvmlite==0.41.1
MarkupSafe==2.1.3
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
networkx==3.1
numba==0.58.1
numpy==1.25.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pansi==2020.7.3
pefile==2023.2.7
peppercorn==0.6
pipreqs==0.4.13
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.24.2
py2neo==2021.2.4
pycparser==2.21
Pygments==2.17.2
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqt5-tools==5.15.9.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sklearn==0.0.post7
some-package==0.1
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.32.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarg==0.1.9
